+++ tests/base/test_evaluation.py
@@ -14,13 +14,16 @@ from rasa_nlu.evaluate import (
     remove_empty_intent_examples, get_entity_extractors,
     get_duckling_dimensions, known_duckling_dimensions,
     find_component, remove_duckling_extractors, drop_intents_below_freq,
-    run_cv_evaluation, substitute_labels, IntentEvaluationResult)
+    run_cv_evaluation, substitute_labels, IntentEvaluationResult,
+    evaluate_intents)
 from rasa_nlu.evaluate import does_token_cross_borders
 from rasa_nlu.evaluate import align_entity_predictions
 from rasa_nlu.evaluate import determine_intersection
 from rasa_nlu.evaluate import determine_token_labels
 from rasa_nlu.config import RasaNLUModelConfig
 from rasa_nlu.tokenizers import Token
+from rasa_nlu import utils
+import json
 from rasa_nlu import training_data, config
 from tests import utilities
 
@@ -256,6 +259,41 @@ def test_run_cv_evaluation():
     assert len(entity_results.test['ner_crf']["Accuracy"]) == n_folds
     assert len(entity_results.test['ner_crf']["Precision"]) == n_folds
     assert len(entity_results.test['ner_crf']["F1-score"]) == n_folds
+
+
+def test_evaluation_report(tmpdir_factory):
+
+    path = tmpdir_factory.mktemp("evaluation").strpath
+    report_filename = path + "report.json"
+
+    intent_results = [
+        IntentEvaluationResult("", "restaurant_search",
+                               "I am hungry", 0.12345),
+        IntentEvaluationResult("greet", "greet",
+                               "hello", 0.98765)]
+
+    result = evaluate_intents(intent_results,
+                              report_filename,
+                              successes_filename=None,
+                              errors_filename=None,
+                              confmat_filename=None,
+                              intent_hist_filename=None)
+
+    report = json.loads(utils.read_file(report_filename))
+
+    greet_results = {"precision": 1.0,
+                     "recall": 1.0,
+                     "f1-score": 1.0,
+                     "support": 1}
+
+    prediction = {'text': 'hello',
+                  'intent': 'greet',
+                  'predicted': 'greet',
+                  'confidence': 0.98765}
+
+    assert len(report.keys()) == 4
+    assert report["greet"] == greet_results
+    assert result["predictions"][0] == prediction
 
 
 def test_empty_intent_removal():